+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix regex validation of VMware VMM vCenter username to allow for `domain\username` format
 - Add health score evaluation policy
 - Add support for enhanced LAGs and uplink configuration for VMware VMM domains
+- Add fabric SPAN destination group
 
 ## 0.3.4